Adds ability to see and search moves
