+++ src/app/store.ts
@@ -1,9 +1,11 @@
 import { configureStore, ThunkAction, Action } from '@reduxjs/toolkit';
-import pokedexReducer from '../features/pokedex/pokedexSlice';
+import pokedexReducer from '../features/pokedex/slices/pokedexSlice';
+import pokemonReducer from '../features/pokedex/slices/pokemonSlice';
 
 export const store = configureStore({
   reducer: {
     pokedex: pokedexReducer,
+    pokemon: pokemonReducer
   },
 });
