+++ src/features/pokedex/Pokedex.tsx
@@ -1,14 +1,17 @@
 import { useState } from "react";
-import { pokedexSearch, selectHistory, selectStatus } from "./pokedexSlice";
+import { pokedexSearch, selectHistory, selectStatus } from "./slices/pokedexSlice";
 import { useAppDispatch, useAppSelector } from '../../app/hooks';
 import { Box, Form, Text, TextInput, Button } from 'grommet';
-import { PokedexCard } from "./PokedexCard";
-import { PokedexHistory } from "./PokedexHistory";
+import { PokedexCard } from "./components/PokedexCard";
+import { PokedexHistory } from "./components/PokedexHistory";
+import { selectMove } from "./slices/pokemonSlice";
 
 export function Pokedex() {
   const dispatch = useAppDispatch();
   const searchStatus = useAppSelector(selectStatus);
   const historyStatus = useAppSelector(selectHistory);
+  const moveDescription = useAppSelector(selectMove);
+
   const [searchTerm, setSearchTerm] = useState<string>('');
 
   function handleSearch(){
@@ -39,10 +42,12 @@ export function Pokedex() {
         </Form>
       </Box>
       {searchStatus === 'fulfilled' && <PokedexCard/>}
+      <Text alignSelf='center' color="green">{moveDescription}</Text>
       {searchStatus === 'failed' && <Text color="red" alignSelf="center">No Pokemon with that name was found. Please try your search again.</Text>}
       <Box flex justify="center" direction="row">
         {historyStatus.length > 0 && <PokedexHistory/>}
       </Box>
+
     </Box>
   );
 }
